COT-48 CAutomate::QMap<QString, IVariable*>m_mapStreams -> CAutomoate::QList<CVariableStream*>m_listStreams Quelques modifs dans la partie gui pour que ça compile à vérifier.

+++ libcot/automate/cycles/CCycleIO.h
@@ -18,7 +18,7 @@ public slots:
     void slotRunCycle() Q_DECL_OVERRIDE;
     void slotPauseCycle() Q_DECL_OVERRIDE;
     void slotStopCycle() Q_DECL_OVERRIDE;
-    void addAction(int, IAction*) Q_DECL_OVERRIDE;
+    void addAction(float, IAction*) Q_DECL_OVERRIDE;
 
 
 private:
